pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-mlops-assignment-04"
version = "0.1.0"
description = "End-to-end MLOps assignment for IEEE-CIS fraud detection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.135.1",
  "joblib==1.4.2",
  "kfp==2.8.0",
  "lightgbm==4.6.0",
  "numpy==2.1.1",
  "pandas==2.2.3",
  "prometheus-client==0.25.0",
  "pydantic>=2.8.0",
  "pyyaml==6.0.2",
  "scikit-learn==1.5.2",
  "scipy>=1.13.0",
  "shap==0.51.0",
  "uvicorn>=0.30.0",
  "xgboost==3.2.0",
  "imbalanced-learn==0.14.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
